requirements.txt
streamlit==1.28.1
openai==1.3.3
beautifulsoup4==4.12.2
pandas
requests
Pillow
